backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
ecdsa==0.19.2
fastapi==0.135.3
filelock==3.28.0
firecrawl-py==4.22.1
fonttools==4.62.1
fsspec==2026.3.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-aiplatform==1.148.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.4.0
groq==1.1.2
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
nest-asyncio==1.6.0
packaging==26.1
passlib==1.7.4
piexif==1.1.3
pillow==12.2.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
requests==2.33.1
rsa==4.9.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tavily-python==0.7.23
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
websockets==16.0
yarl==1.23.0
h3==4.4.2
stripe==15.0.1
kafka-python==2.3.1